Backport extra_float_digits GUC maximum increase
We need this commit for gpbackup to backup float4 values correctly.
Without the extra precision, float4 columns that are being used as the
distribution key may not be backed up correctly for the distribution
hash check during COPY FROM SEGMENT.

As part of this backport, we also changed the logic in cdb_dump which is
Greenplum code.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Jimmy Yih <jyih@pivotal.io>

commit 8c5463a51176c8d2a01fcf154d7ac33fa9a74f6a
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Fri Sep 11 19:17:04 2009 +0000

    Increase the maximum value of extra_float_digits to 3, and have pg_dump
    use that value when the backend is new enough to allow it.  This responds
    to bug report from Keh-Cheng Chu pointing out that although 2 extra digits
    should be sufficient to dump and restore float8 exactly, it is possible to
    need 3 extra digits for float4 values.