Conditionally set $PYTHONHOME only when vendoring python
Setting `$PYTHONHOME` for releases where python is not vendored causes
the system python to fail with an error message about not being able to
find the site module.

This commit updates the logic in generate-greenplum_path.sh to check if
we are vendoring python and if we are, it includes setting `$PYTHONHOME`
in greenplum_path.sh. If we are vendoring python, then at build-time
`$PYTHONHOME` is set to point at our vendored python.

[#173046174]

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>