Fix greenplum_path script for OSS Ubuntu
We do not vendor python on OSS Ubuntu 16.04 and therefore
`$GPHOME/ext/python` doesn't exist. This causes the system python to
fail with an error message about not being able to find the site module.

This commit updates the logic in generate-greenplum_path.sh to check if
we are vendoring python and if we are, it includes setting `$PYTHONHOME`
in greenplum_path.sh. If we are vendoring python, then at build-time
`$PYTHONHOME` is set to point at our vendored python.

[#173046174]

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>