Use getpass.getuser() to determine login name of the user.
https://docs.python.org/3.7/library/getpass.html#getpass.getuser

Return the “login name” of the user.

This function checks the environment variables LOGNAME, USER, LNAME and USERNAME, in order, and returns the value of the first one which is set to a non-empty string. If none are set, the login name from the password database is returned on systems which support the pwd module, otherwise, an exception is raised.

In general, this function should be preferred over os.getlogin().

Authored-by: Brent Doil <bdoil@vmware.com>